ModIcon refactor
